enrich_job_listings/requirements.txt
beautifulsoup4==4.12.3
fake_useragent==1.5.1
pandas==2.2.1
pandas_gbq==0.23.1
Requests==2.32.3
google-auth==2.29.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0